requirements.txt
gunicorn
flask
airtable-python-wrapper
pandas==1.2.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
google-spreadsheet==0.0.6
googleapis-common-protos==1.53.0
pytz==2021.1